Handle all quoting besides Here-Documents
This change begins to refactor the lexer into a more object oriented
approach. However, it takes on some technical debt in the process. There
is also an egregious hack to make sure put backslashes into the output
when inside double quotes. The reason is that to determine if the
backslash should appear in the output requires looking ahead. I got
around this with a check at the start of the main lex loop.